Move Image Annotation file and remove old one
Commit removes Image Annotation file and removes the old file with Annotorious version 2
